Gate document generation behind ticket claim, fix duplicate save log, more polish
- PIC picker (job creation form + Change Current Responsible) now
  lists all staff across every department instead of filtering to
  the job's own department — a staff member can be assigned outside
  their usual department.
- Every date display across the app (Deadline, Mula, installment due
  dates, Settings) now spells out the full month name and year,
  matching the Activity Log timestamp fix from the last round.
- Clicking Simpan in a document preview with no actual changes no
  longer writes a duplicate "dikemaskini" entry to the Activity Log.
- Documents can't be generated (and items can't be edited) until a
  ticket has actually been claimed — the Dokumen buttons are disabled
  with an explanatory note until Take In Ticket / Change Current
  Responsible sets a PIC, so nobody can quietly bill a ticket that
  isn't officially anyone's yet.
- "Take In Ticket" only shows in the Action menu while a ticket is
  still unclaimed; once it has a PIC, only "Change Current
  Responsible" makes sense there.
- Finance's Ledger table gains Debit/Credit account columns (using
  the existing ledgerAccountLabel formatter) instead of only showing
  the net amount per entry.

+++ app/settings/page.jsx
@@ -2,7 +2,7 @@
 import { useState, useMemo, useEffect } from "react";
 import { useAuth, useData } from '@/lib/hooks';
 import { DEPT, ROLE } from '@/lib/constants';
-const formatDate=d=>d?new Date(d).toLocaleDateString("ms-MY",{day:"numeric",month:"short",year:"numeric"}):"—";
+const formatDate=d=>d?new Date(d).toLocaleDateString("ms-MY",{day:"numeric",month:"long",year:"numeric"}):"—";
 
 function RBadge({r}){const m=ROLE[r];return m?<span className="badge-r" style={{color:m.color,background:m.color+"15"}}>{m.label}</span>:null}
 function DTag({d}){if(!d)return<span className="text-sm text-muted">Semua</span>;const m=DEPT[d];return m?<span className="badge-d" style={{color:m.color,background:m.color+"15"}}>{m.label}</span>:null}
